% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccbingo.R, R/cards.R
\docType{data}
\name{ccbingo}
\alias{ccbingo}
\alias{char_to_seed}
\alias{ccbingo_app}
\alias{cards}
\title{Conference call bingo}
\format{
\code{cards} is a \emph{named} list (by topic) of strings to use for cards.
}
\usage{
ccbingo(card, seed = 1L, wins = c("col", "row", "corner", "diag"), ...)

char_to_seed(player)

ccbingo_app()

cards
}
\arguments{
\item{card}{a vector of character strings to make random cards; strings
are randomly selected from \code{card} without replacement but recycled
as needed to draw a 5x5 grid}

\item{seed}{an integer passed to \code{\link{set.seed}} to reproduce cards;
alternatively, a character string to be converted to a seed with
\code{char_to_seed}}

\item{wins}{ways to win the game, any combination of \code{"col"},
\code{"row"}, \code{"corner"}, or \code{"diag"}; alternatively, a function
to determine wins returning \code{''} for no win and a non-zero length
string for wins}

\item{...}{additional arguments passed to \code{ccbingo:::bingo} such as
\code{title} (card title), \code{free} (free space text), \code{width}
(text width per box), \code{col.grid} (grid colors), \code{col.select}
(color of selected boxes), or \code{col.free} (color of free box)}

\item{player}{player name}
}
\description{
Play conference call bingo locally or with \code{\link[shiny]{shiny}}.

\code{ccbingo} runs in a regular \code{R} session and prompts the user for
each click using \code{\link{locator}}. \code{ccbingo:::ccbingo_shiny} has
similar functionality but requires a list of click coordinates.

\code{char_to_seed} creates a simple string-to-seed mapping which allows
cards to be reproduced from a player name, for example.

\code{cards} is a list of built-in phrases to use for cards.
}
\examples{
char_to_seed('joe')
char_to_seed('sue')

\dontrun{
## these are not exported
ccbingo:::bingo()
ccbingo:::bingo(letters)
ccbingo:::bingo(
  cards$zoom, col.grid = rainbow(6), col.free = 'white',
  free = '<3', title = 'come and play'
)

## play interactively in an r session
ccbingo(cards$zoom)
ccbingo(cards$zoom, title = 'my bingo', free = '<3', col.free = 'white',
        col.select = 'transparent', col.grid = rainbow(6))


## same as above except pass a list of recorded clicks

## interactive -- first get up a game then make clicks
# ccbingo(cards$zoom)
# clicks <- locator(5)
# ccbingo:::ccbingo_shiny(cards$zoom, clicks = clicks)

## or pass a list of clicks directly to game
set.seed(1)
clicks <- list(x = runif(15, 0, 5), y = -runif(15, 0, 5))
ccbingo:::ccbingo_shiny(cards$zoom, clicks = clicks)


## run shiny app
ccbingo_app()
}

}
\keyword{datasets}
